perf: eliminated useless packages + lazy loading components + restructured animations
