+++ package.json
@@ -30,7 +30,6 @@
 		"@tailwindcss/nesting": "0.0.0-insiders.565cd3e",
 		"@xata.io/client": "^0.28.4",
 		"gsap": "^3.12.5",
-		"luxon": "^3.4.4",
 		"react": "^18.2.0",
 		"react-email": "1.9.5",
 		"rehype-preset-minify": "^7.0.0",
@@ -81,7 +80,6 @@
 		"puppeteer": "^21.9.0",
 		"svelte": "^3.59.2",
 		"svelte-check": "^3.6.3",
-		"svelte-french-toast": "^1.2.0",
 		"tailwindcss": "^3.4.1",
 		"tsafe": "^1.6.6",
 		"tslib": "^2.6.2",
