+++ src/routes/(other)/+layout.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-	import { Toaster } from "svelte-french-toast";
-
 	import Footer from "./Footer.svelte";
 	import Header from "./Header.svelte";
 </script>
@@ -16,5 +14,3 @@
 </main>
 
 <Footer />
-
-<Toaster />
